requirements.txt
attrs==25.3.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dotenv==1.1.1
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
urllib3==2.5.0
